feat: Adding openai abstraction and first prompt

+++ src/main/java/br/com/fiap/models/entities/CandidateBehaviorProfile.java
@@ -3,11 +3,11 @@ package br.com.fiap.models.entities;
 import java.time.LocalDateTime;
 
 public class CandidateBehaviorProfile {
-    private int candidateId;
+    private Long candidateId;
     private String aiProfile;
     private LocalDateTime lastUpdated;
 
-    public CandidateBehaviorProfile(int candidateId, String aiProfile) {
+    public CandidateBehaviorProfile(Long candidateId, String aiProfile) {
         this.candidateId = candidateId;
         this.aiProfile = aiProfile;
         this.lastUpdated = LocalDateTime.now();
@@ -17,8 +17,8 @@ public class CandidateBehaviorProfile {
         this.lastUpdated = LocalDateTime.now();
     }
 
-    public int getCandidateId() { return candidateId; }
-    public void setCandidateId(int candidateId) { this.candidateId = candidateId; }
+    public Long getCandidateId() { return candidateId; }
+    public void setCandidateId(Long candidateId) { this.candidateId = candidateId; }
 
     public String getAiProfile() { return aiProfile; }
     public void setAiProfile(String aiProfile) { this.aiProfile = aiProfile; }
